Force serviceEmailaddress to contact me form
On the runner the service#serviceEmailAddress is underfined,
so the quick fix is to force the config.

When the runner get the instance property

```
email.address.from
propDefaultValue: undefined
propValue: "[% service#name %]" <[% service#serviceEmailAddress %]>
```

The serviceEmailAddress is blank on the runner context but does exist
on the config.json in the fb-components.